+++ src/app/Components/icons/LikeButton/index.tsx
@@ -1,80 +1,48 @@
-"use client";
-
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { GoHeart, GoHeartFill } from "react-icons/go";
-import { UseFirestore } from "@/app/Context/FirestoreContext";
+import { useLikes } from "@/app/Context/LikesContext";
 import { UseAuth } from "@/app/Context/AuthContext";
-import { db } from "@/app/firebase";
-import getFirebaseDocumentId from "../../FirebaseDocumentId";
-import { doc, getDoc, updateDoc } from "firebase/firestore";
 import LogInWarningModal from "./LogInWarningModal";
 
-const LikeButton = ({ id }: { id: number }) => {
-  const { chapters } = UseFirestore();
+interface LikeButtonProps {
+  id: number;
+}
+
+const LikeButton: React.FC<LikeButtonProps> = ({ id }) => {
+  const { likes, toggleLike, likeLoadingStates } = useLikes();
   const { user } = UseAuth();
-
-  const email = user?.email;
-  const currentChapter = chapters.find((chapter) => Number(chapter.id) === id);
-
-  const [isLiked, setIsLiked] = useState(
-    currentChapter?.likes?.includes(email)
-  );
-  const [likeCount, setLikeCount] = useState(
-    currentChapter?.likes?.length || 0
-  );
   const [isModalOpen, setIsModalOpen] = useState(false);
 
-  useEffect(() => {
-    setIsLiked(currentChapter?.likes?.includes(email));
-    setLikeCount(currentChapter?.likes?.length || 0);
-  }, [currentChapter, email]);
+  const userLikes = likes[id] || [];
+  const email = user?.email;
+  const isLiked = userLikes.includes(email || "");
+  const likeCount = userLikes.length;
+  const isLikeLoading = likeLoadingStates[id];
 
-  const likeChapter = async () => {
+  const handleLikeClick = async () => {
     if (!email) {
       setIsModalOpen(true);
       return;
     }
-
-    const chapterId = await getFirebaseDocumentId("Chapters", "id", id);
-    const chapterRef = doc(db, "Chapters", chapterId!.toString());
-    const docSnap = await getDoc(chapterRef);
-
-    if (!docSnap.exists()) {
-      console.log("Document does not exist!");
-      return;
-    }
-
-    const chapterData = docSnap.data();
-    const userLikes = chapterData.likes || [];
-    const alreadyLiked = userLikes.includes(email);
-
-    if (alreadyLiked) {
-      const newLikes: string[] = userLikes.filter(
-        (like: string) => like !== email
-      );
-      await updateDoc(chapterRef, { likes: newLikes });
-      setIsLiked(false);
-      setLikeCount(newLikes.length);
-      console.log(`unliked chapter: ${id} by ${email}`);
-    } else {
-      const newLikes = [...userLikes, email];
-      await updateDoc(chapterRef, { likes: newLikes });
-      setIsLiked(true);
-      setLikeCount(newLikes.length);
-      console.log(`liked chapter: ${id} by ${email}`);
-    }
+    await toggleLike(id);
   };
 
   return (
     <div className="flex flex-col items-center justify-center">
-      <button onClick={likeChapter} className="focus:outline-none h-full">
+      <button onClick={handleLikeClick} className="focus:outline-none h-full">
         {isLiked ? (
-          <GoHeartFill className="text-red-500 h-full cursor-pointer" />
+          <GoHeartFill
+            className={`text-red-500 h-full cursor-pointer ${
+              isLikeLoading && "opacity-30"
+            }`}
+          />
         ) : (
-          <GoHeart className="cursor-pointer h-full" />
+          <GoHeart
+            className={`cursor-pointer h-full ${isLikeLoading && "opacity-30"}`}
+          />
         )}
       </button>
-      <p>{likeCount}</p>
+      <p className={`${isLikeLoading && "opacity-30"}`}>{likeCount}</p>
       <LogInWarningModal
         isOpen={isModalOpen}
         onClose={() => setIsModalOpen(false)}